ユーザー削除APIの権限チェックを追加

+++ ApplicationService/Users/UserDeleteApplicationService.php
@@ -5,6 +5,9 @@ namespace App\ApplicationService\Users;
 
 use App\Domain\Models\User\IUserRepository;
 use App\Domain\Models\User\Type\UserId;
+use App\Domain\Services\PermissionService;
+use App\Domain\Shared\Exception\ExceptionItem;
+use App\Domain\Shared\Exception\PermissionDeniedException;
 
 /**
  * class UserDeleteApplicationService
@@ -17,14 +20,22 @@ class UserDeleteApplicationService {
 	private $__userRepository;
 
 /**
+ * @var \App\Domain\Services\PermissionService
+ */
+	private $__permissionService;
+
+/**
  * constructor
  *
  * @param \App\Domain\Models\User\IUserRepository $userRepository userRepository
+ * @param \App\Domain\Services\PermissionService $permissionService permissionService
  */
 	public function __construct(
-		IUserRepository $userRepository
+		IUserRepository $userRepository,
+		PermissionService $permissionService
 	) {
 		$this->__userRepository = $userRepository;
+		$this->__permissionService = $permissionService;
 	}
 
 /**
@@ -32,8 +43,13 @@ class UserDeleteApplicationService {
  *
  * @param \App\ApplicationService\Users\UserDeleteCommand $command command
  * @return void
+ * @throws \App\Domain\Shared\Exception\PermissionDeniedException
  */
 	public function handle(UserDeleteCommand $command) : void {
+		if (!$this->__permissionService->isAllowedSelf('add', 'user')) {
+			throw new PermissionDeniedException([new ExceptionItem('', '権限がありません。')]);
+		}
+
 		$id = new UserId($command->getUserId());
 		$this->__userRepository->delete($id);
 	}
